requirements.txt
mammoth==1.8.0
selenium==4.23.0
python-docx==1.1.2